pyproject.toml
[tool.ruff]
line-length = 90
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "D2",
    "D3",
    "D4",
    "Q",
    "SIM",
    "RUF",
    "ERA",
    "RSE",
    "ASYNC"
]
fixable = ["F","I","D","T20","ERA"]
ignore = ["F403","D213","D203","F405"]

[tool.ruff.lint.per-file-ignores]
"errors.py" = ["E501"]
"errorhandler.py" = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
pythonPlatform = "All"
pythonVersion = "3.12"

[tool.poetry]
name = "bbombsbot"
version = "0.1.0"
description = "Experimental Discord bot written in Python with Hikari."
authors = ["BBombs <133826485+bbombss@users.noreply.github.com>"]
license = "GNU GPL-3.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
hikari = "2.0.0.dev125"
hikari-lightbulb = "~=2.3.5"
asyncpg = "~=0.29.0"
psutil = "^5.9.6"
aiofiles = "^23.2.1"
uvloop = {version="0.19.0", platform="linux"}
hikari-miru = "~=4.1.1"
levenshtein = "^0.25.1"

[tool.poetry.dev-dependencies]
ruff = "^0.1.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
